Splitting the generatorCCFGCompiler file into multiple folders and files for easier readability
