+++ packages/TCOS/src/language-server/structural-operational-semantics-module.ts
@@ -11,19 +11,8 @@ import { registerValidationChecks } from '../../node_modules/langium/lib/grammar
 import { StructuralOperationalSemanticsGeneratedModule, StructuralOperationalSemanticsGeneratedSharedModule } from './generated/module.js';
 import { StructuralOperationalSemanticsValidator, registerSoSValidationChecks } from './structural-operational-semantics-validator.js';
 import { SoSScopeProvider } from './sos-scope.js';
-// import { LangiumGrammarDocument } from 'langium/lib/grammar/workspace/documents';
 import { registerTypeValidationChecks } from '../../node_modules/langium/lib/grammar/validation/types-validator.js';
 import { SoSSemanticTokenProvider } from './structural-operational-semantics-semantic-token.js';
-
-
-//import {LangiumGrammarGeneratedModule /*, LangiumGrammarGeneratedSharedModule*/ } from 'langium/src/grammar/generated/module';
-//import { LangiumGrammarModule} from 'langium';
-
-
-//import { SoSScopeComputation, SoSScopeProvider } from './sos-scope';
-
-
-
 
 /**
  * Declaration of custom services - add your own service classes here.
@@ -46,12 +35,7 @@ export type StructuralOperationalSemanticsServices = LangiumServices & Structura
  * selected services, while the custom services must be fully specified.
  */
 export const StructuralOperationalSemanticsModule: Module<StructuralOperationalSemanticsServices, PartialLangiumServices & StructuralOperationalSemanticsAddedServices> = {
-    references: {
-  //      ScopeComputation: (services) => new SoSScopeComputation(services),
-        ScopeProvider: (services) => new SoSScopeProvider(services)//,
-        // QualifiedNameProvider: () => new QualifiedNameProvider()
-    },
-    
+    references: { ScopeProvider: (services) => new SoSScopeProvider(services) },
     validation: {
         StructuralOperationalSemanticsValidator: () => new StructuralOperationalSemanticsValidator()
     },
@@ -89,15 +73,8 @@ export function createStructuralOperationalSemanticsServices(context: DefaultSha
         StructuralOperationalSemanticsGeneratedModule,
         StructuralOperationalSemanticsModule
     );
-    const langiumServices = /*inject(
-        createDefaultModule({ shared }),
-        LangiumGrammarGeneratedModule,
-        LangiumGrammarModule
-    );*/
-    createLangiumGrammarServices(context).grammar
+    const langiumServices = createLangiumGrammarServices(context).grammar
 
-
-    // addTypeCollectionPhase(shared, langiumServices);
     shared.ServiceRegistry.register(langiumServices);
 
     registerValidationChecks(langiumServices);
@@ -107,14 +84,3 @@ export function createStructuralOperationalSemanticsServices(context: DefaultSha
     registerSoSValidationChecks(StructuralOperationalSemantics);
     return { shared,langiumServices, StructuralOperationalSemantics };
 }
-// function addTypeCollectionPhase(sharedServices: LangiumSharedServices, grammarServices: LangiumGrammarServices) {
-//     const documentBuilder = sharedServices.workspace.DocumentBuilder;
-//     documentBuilder.onBuildPhase(DocumentState.IndexedReferences, async (documents, cancelToken) => {
-//         for (const document of documents) {
-//             await interruptAndCheck(cancelToken);
-//             const typeCollector = grammarServices.validation.ValidationResourcesCollector;
-//             const grammar = document.parseResult.value as Grammar;
-//             (document as LangiumGrammarDocument).validationResources = typeCollector.collectValidationResources(grammar);
-//         }
-//     });
-// }
