+++ packages/TCOS/src/language-server/structural-operational-semantics-validator.ts
@@ -37,14 +37,4 @@ export class StructuralOperationalSemanticsValidator {
             });
         }
     }
-
-    // checkPersonStartsWithCapital(person: Person, accept: ValidationAcceptor): void {
-    //     if (person.name) {
-    //         const firstChar = person.name.substring(0, 1);
-    //         if (firstChar.toUpperCase() !== firstChar) {
-    //             accept('warning', 'Person name should start with a capital.', { node: person, property: 'name' });
-    //         }
-    //     }
-    // }
-
 }
