+++ packages/TCOS/tsconfig.json
@@ -4,7 +4,7 @@
     "target": "ES2021",
     "module": "Node16",
     "lib": ["ESNext"],
-    //"rootDir": "./src",
+    "rootDir": "./src",
     "outDir": "out/",
     "strict": true,
     "noUnusedLocals": true,
